First Commit
Connect4 Game: A Java-based implementation of the classic Connect4 Four game with a modern twist. Built using JavaFX for the graphical user interface and XML for data storage, this project offers a seamless and visually engaging gaming experience. Challenge your friends or test your strategic skills against the computer in this fun and interactive rendition of the timeless Connect4 game.